feat: Improve build configs and add centralized scripts
