+++ packages/build-tooling/scripts/build-rollup.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+rollup -c rollup.config.ts --configPlugin typescript && pnpm types:build
