+++ packages/build-tooling/scripts/build-vite.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+vite build && pnpm types:build
