+++ packages/build-tooling/scripts/format.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+pnpm prettier --write --ignore-path=../../.prettierignore .
