+++ packages/build-tooling/scripts/types-build-vue.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+vue-tsc -p tsconfig.build.json && tsc-alias -p tsconfig.build.json
