+++ packages/build-tooling/scripts/types-build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tsc -p tsconfig.build.json && tsc-alias -p tsconfig.build.json
