+++ packages/build-tooling/scripts/types-check-vue.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+vue-tsc --noEmit
