+++ packages/build-tooling/scripts/types-check.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tsc --noEmit
