+++ packages/build-tooling/src/index.ts
@@ -1,6 +1,3 @@
-export { findEntryPoints } from './entry'
-export {
-  createRollupConfig,
-  createViteBuildOptions,
-  autoCSSInject,
-} from './build-options'
+export { findEntryPoints, addPackageFileExports } from './entry'
+export { createRollupConfig } from './rollup-options'
+export { alias, createViteBuildOptions, autoCSSInject } from './vite-options'
