+++ packages/build-tooling/tsconfig.build.json
@@ -1,20 +1,11 @@
 {
   "extends": "./tsconfig.json",
   "include": ["src"],
-  "exclude": [
-    "dist",
-    "test",
-    "vite.config.ts",
-    "**/*.test.ts",
-    "**/*.test.json",
-    "**/*.test.md"
-  ],
   "compilerOptions": {
     "skipLibCheck": true,
     "declaration": true,
     "declarationMap": true,
     "emitDeclarationOnly": true,
-    "outDir": "dist/",
-    "target": "ESNext"
+    "outDir": "dist/"
   }
 }
