+++ packages/code-highlight/package.json
@@ -21,14 +21,14 @@
     "node": ">=18"
   },
   "scripts": {
-    "build": "vite build && pnpm types:build && tsc-alias -p tsconfig.build.json",
+    "build": "scalar-build-rollup",
     "dev": "vite",
+    "format": "scalar-format-js",
     "lint:check": "eslint .",
     "lint:fix": "eslint .  --fix",
-    "preview": "vite preview",
     "test": "vitest",
-    "types:build": "tsc -p tsconfig.build.json",
-    "types:check": "tsc --noEmit --skipLibCheck --composite false"
+    "types:build": "scalar-types-build",
+    "types:check": "scalar-types-check"
   },
   "type": "module",
   "main": "dist/index.js",
@@ -36,6 +36,18 @@
     ".": {
       "import": "./dist/index.js",
       "types": "./dist/index.d.ts"
+    },
+    "./rehype-highlight": {
+      "import": "./dist/rehype-highlight/index.js",
+      "types": "./dist/rehype-highlight/index.d.ts"
+    },
+    "./markdown": {
+      "import": "./dist/markdown/index.js",
+      "types": "./dist/markdown/index.d.ts"
+    },
+    "./code": {
+      "import": "./dist/code/index.js",
+      "types": "./dist/code/index.d.ts"
     },
     "./css/*.css": {
       "import": "./dist/css/*.css",
