+++ packages/code-highlight/src/code/index.ts
@@ -1,0 +1,1 @@
+export { syntaxHighlight } from './highlight'
