+++ packages/code-highlight/src/index.ts
@@ -1,3 +1,4 @@
 export { rehypeHighlight } from './rehype-highlight'
 export { lowlightLanguageMappings } from './constants'
 export { htmlFromMarkdown } from './markdown'
+export { syntaxHighlight } from './code'
