+++ packages/code-highlight/src/markdown/index.ts
@@ -1,0 +1,1 @@
+export { htmlFromMarkdown } from './markdown'
