+++ packages/code-highlight/src/rehype-highlight/index.ts
@@ -1,0 +1,1 @@
+export { rehypeHighlight } from './rehype-highlight'
