+++ packages/code-highlight/vite.config.ts
@@ -1,4 +1,4 @@
-import { findEntryPoints } from '@scalar/build-tooling'
+import { alias, findEntryPoints } from '@scalar/build-tooling'
 import { URL, fileURLToPath } from 'node:url'
 import { defineConfig } from 'vite'
 
@@ -7,33 +7,11 @@ import pkg from './package.json'
 export default defineConfig({
   plugins: [],
   resolve: {
-    alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url)),
-    },
+    alias: alias(import.meta.url),
     dedupe: ['vue'],
   },
   server: {
     port: 9000,
   },
   publicDir: './src/css',
-  build: {
-    ssr: true,
-
-    minify: false,
-    target: 'esnext',
-    lib: {
-      entry: await findEntryPoints({ allowCss: true }),
-      formats: ['es'],
-    },
-    rollupOptions: {
-      external: [...Object.keys(pkg.dependencies)],
-      output: {
-        // Create a separate file for the dependency bundle
-        manualChunks: (id) =>
-          id.includes('node_modules') ? 'vendor' : undefined,
-        format: 'esm',
-      },
-      treeshake: true,
-    },
-  },
 })
